Fix an issue with nesting functions that support ShouldProcess
By calling the outer function with a `-Confirm:$false` that then becomes part of the PSBoundParameters. Then when we splat that on the inner function and add our own `-Confirm:$false` PowerShell gets upset because the same parameter is provided twice. I believe this is being fixed in PS 7.2 but that doesn't do a library type module like this much good.